requirements.txt
Django==4.2.7
django_imagekit==4.1.0
djangorestframework==3.14.0
drf_spectacular==0.26.4
